requirements.txt
pillow~=11.0.0
pyzbar~=0.1.9
cryptography~=44.0.2
opencv-python~=4.11.0.86
docopt~=0.6.2
numpy~=2.2.4
matplotlib~=3.10.1
tqdm~=4.67.1
rich~=13.9.4
requests~=2.32.3
crypto~=1.4.1
sympy~=1.13.1
rsa~=4.9
colorama~=0.4.6